+++ operators/advanced_network/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Holoscan Advanced Network Operator Changelog
+# Holoscan Advanced Network library Changelog
 
 ## 🚧 holoscan-networking 0.2
 
@@ -7,10 +7,18 @@
 ### Breaking
 
 - Removed `get_port_from_ifname`. Use `address_to_port` instead.
+- Removed the `adv_network_tx` (`AdvNetworkOpTx`) and `adv_network_rx` operators. See https://github.com/nvidia-holoscan/holohub/pull/743. Instead:
+  - Initialize your NIC with `advanced_network::adv_net_init` in your application.
+  - In your custom Tx operator, replace `op_output.emit` with `advanced_network::send_tx_burst`.
+  - In your custom Rx operator, replace `op_input.receive` with `advanced_network::get_rx_burst`, also specifying the port and queue ids.
 
 ### Non-Breaking
 
--
+- Common:
+  - `address_to_port` now also supports interface names as input, not just pcie addresses.
+  - Added `get_num_rx_queues` to get the number of RX queues for a given port.
+
+
 
 ## 🚀 holoscan-networking 0.1
 
@@ -32,7 +40,6 @@
 - Common:
   - Switch to flow_isolation mode by default. See https://github.com/nvidia-holoscan/holohub/pull/624.
   - Fixed lockup with low buffer_size. See https://github.com/nvidia-holoscan/holohub/pull/735.
-- DPDK manager:
   - Support multiple queues per core. See https://github.com/nvidia-holoscan/holohub/pull/638.
   - Avoid crash on DPDK init failures. See https://github.com/nvidia-holoscan/holohub/pull/725.
   - Added stat reporting. See https://github.com/nvidia-holoscan/holohub/pull/735.